support font8x8

+++ bwapi/include/BWAPI/Interface.h
@@ -69,7 +69,7 @@ namespace BWAPI
     template <typename CT>
     CT getClientInfo(int key = 0) const
     {
-      return (CT)(int)this->getClientInfo(key);
+      return (CT)(long)this->getClientInfo(key);
     };
 
     /// <summary>Associates one or more pointers or values with any BWAPI interface.</summary>
